+++ projects/packages/jetpack-mu-wpcom/src/features/verbum-comments/src/components/types.d.ts
@@ -6,6 +6,14 @@ type ScriptLoader = {
 
 declare global {
 	const VerbumComments: VerbumCommentsType;
+	const verbumBlockEditor: {
+		attachGutenberg: (
+			textarea: HTMLTextAreaElement,
+			content: ( embedUrl: string ) => void,
+			isRtl: boolean,
+			onEmbedContent: ( embedUrl: string ) => void
+		) => void;
+	};
 	const vbeCacheBuster: string;
 	const WP_Enqueue_Dynamic_Script: ScriptLoader;
 	const wp: Record< string, unknown >;