+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-admin-interface/wpcom-admin-interface.php
@@ -119,6 +119,7 @@ add_filter( 'pre_update_option_wpcom_admin_interface', 'wpcom_admin_interface_pr
 
 const WPCOM_DUPLICATED_VIEW = array(
 	'edit.php',
+	'edit.php?post_type=page',
 	'edit.php?post_type=jetpack-portfolio',
 	'edit.php?post_type=jetpack-testimonial',
 	'edit-comments.php',
@@ -392,7 +393,8 @@ add_action( 'admin_notices', 'wpcom_show_admin_interface_notice' );
  */
 function wpcom_is_duplicate_views_experiment_enabled() {
 	$experiment_platform = 'calypso';
-	$experiment_name     = "{$experiment_platform}_post_onboarding_holdout_120924";
+	$experiment_name     = "{$experiment_platform}_post_onboarding_holdout_160125";
+	$aa_test_name        = "{$experiment_platform}_post_onboarding_aa_150125";
 
 	static $is_enabled = null;
 	if ( $is_enabled !== null ) {
@@ -400,11 +402,12 @@ function wpcom_is_duplicate_views_experiment_enabled() {
 	}
 
 	if ( ( new Host() )->is_wpcom_simple() ) {
+		\ExPlat\assign_current_user( $aa_test_name );
 		$is_enabled = 'treatment' === \ExPlat\assign_current_user( $experiment_name );
 		return $is_enabled;
 	}
 
-	$option_name = 'duplicate_views_experiment_assignment';
+	$option_name = 'remove_duplicate_views_experiment_assignment';
 	$variation   = get_user_option( $option_name, get_current_user_id() );
 
 	if ( false !== $variation ) {
@@ -416,6 +419,12 @@ function wpcom_is_duplicate_views_experiment_enabled() {
 		$is_enabled = false;
 		return $is_enabled;
 	}
+
+	$aa_test_request_path = add_query_arg(
+		array( 'experiment_name' => $aa_test_name ),
+		"/experiments/0.1.0/assignments/{$experiment_platform}"
+	);
+	Client::wpcom_json_api_request_as_user( $aa_test_request_path, 'v2' );
 
 	$request_path = add_query_arg(
 		array( 'experiment_name' => $experiment_name ),
@@ -454,6 +463,22 @@ function wpcom_is_duplicate_views_experiment_enabled() {
  * the first time.
  */
 function wpcom_show_removed_calypso_screen_notice() {
+	if ( defined( 'IS_WPCOM' ) && IS_WPCOM ) {
+		$blog_id = get_current_blog_id();
+	} else {
+		$jetpack_options = get_option( 'jetpack_options' );
+		if ( is_array( $jetpack_options ) && isset( $jetpack_options['id'] ) ) {
+			$blog_id = (int) $jetpack_options['id'];
+		} else {
+			$blog_id = get_current_blog_id();
+		}
+	}
+
+	// Do not show notice on sites created after the experiment started (2025-01-16).
+	if ( $blog_id > 240790000 ) { // 240790000 is the ID of a site created on 2025-01-16.
+		return;
+	}
+
 	$admin_menu_class = wpcom_get_custom_admin_menu_class();
 	if ( ! $admin_menu_class ) {
 		return;
@@ -465,9 +490,51 @@ function wpcom_show_removed_calypso_screen_notice() {
 		return;
 	}
 
-	$dismissed_notices = get_user_option( 'wpcom_removed_calypso_screen_dismissed_notices' );
-	if ( is_array( $dismissed_notices ) && in_array( $current_screen, $dismissed_notices, true ) ) {
-		return;
+	if ( ( new Host() )->is_wpcom_simple() ) {
+		$preferences  = get_user_attribute( get_current_user_id(), 'calypso_preferences' );
+		$is_dismissed = $preferences[ 'removed-calypso-screen-dismissed-notice-' . $current_screen ] ?? false;
+		if ( $is_dismissed ) {
+			return;
+		}
+	} else {
+		$notices_dismissed_locally = get_user_option( 'wpcom_removed_calypso_screen_dismissed_notices' );
+		if ( ! is_array( $notices_dismissed_locally ) ) {
+			$notices_dismissed_locally = array();
+		}
+
+		if ( in_array( $current_screen, $notices_dismissed_locally, true ) ) {
+			return;
+		}
+
+		if ( ! ( new Jetpack_Connection() )->is_user_connected() ) {
+			return;
+		}
+
+		$response = Client::wpcom_json_api_request_as_user( '/me/preferences', 'v2' );
+		if ( is_wp_error( $response ) ) {
+			return;
+		}
+
+		$response_code = wp_remote_retrieve_response_code( $response );
+		if ( 200 !== $response_code ) {
+			return;
+		}
+
+		$notices_dismissed_globally = array();
+		$preferences                = json_decode( wp_remote_retrieve_body( $response ), true );
+		foreach ( $preferences as $key => $value ) {
+			if ( $value && preg_match( '/^removed-calypso-screen-dismissed-notice-(.+)$/', $key, $matches ) ) {
+				$notices_dismissed_globally[] = $matches[1];
+			}
+		}
+
+		if ( array_diff( $notices_dismissed_globally, $notices_dismissed_locally ) ) {
+			update_user_option( get_current_user_id(), 'wpcom_removed_calypso_screen_dismissed_notices', $notices_dismissed_globally, true );
+		}
+
+		if ( in_array( $current_screen, $notices_dismissed_globally, true ) ) {
+			return;
+		}
 	}
 
 	if ( ! wpcom_is_duplicate_views_experiment_enabled() ) {
@@ -496,7 +563,6 @@ function wpcom_show_removed_calypso_screen_notice() {
 	$clean_title = trim( $clean_title );
 	$config      = wp_json_encode(
 		array(
-			'imageUrl'     => plugins_url( 'screens/' . sanitize_title( $current_screen ) . '.webp', __FILE__ ),
 			'title'        => $clean_title,
 			'screen'       => $current_screen,
 			'ajaxUrl'      => admin_url( 'admin-ajax.php' ),
@@ -531,18 +597,32 @@ function wpcom_get_custom_admin_menu_class() {
 }
 
 /**
- * Handles the AJAX request to dismiss a notice of a removed Calypsos screen.
+ * Handles the AJAX request to dismiss a notice of a removed Calypso screen.
  */
 function wpcom_dismiss_removed_calypso_screen_notice() {
 	check_ajax_referer( 'wpcom_dismiss_removed_calypso_screen_notice' );
 	if ( isset( $_REQUEST['screen'] ) ) {
-		$screen            = sanitize_text_field( wp_unslash( $_REQUEST['screen'] ) );
-		$dismissed_notices = get_user_option( 'wpcom_removed_calypso_screen_dismissed_notices' );
-		if ( ! is_array( $dismissed_notices ) ) {
-			$dismissed_notices = array();
+		$screen = sanitize_text_field( wp_unslash( $_REQUEST['screen'] ) );
+		if ( ( new Host() )->is_wpcom_simple() ) {
+			$preferences = get_user_attribute( get_current_user_id(), 'calypso_preferences' );
+			$preferences[ 'removed-calypso-screen-dismissed-notice-' . $screen ] = true;
+			update_user_attribute( get_current_user_id(), 'calypso_preferences', $preferences );
+		} else {
+			Client::wpcom_json_api_request_as_user(
+				'/me/preferences',
+				'2',
+				array(
+					'method' => 'POST',
+				),
+				array( 'calypso_preferences' => (object) array( 'removed-calypso-screen-dismissed-notice-' . $screen => true ) )
+			);
+			$notices_dismissed_locally = get_user_option( 'wpcom_removed_calypso_screen_dismissed_notices' );
+			if ( ! is_array( $notices_dismissed_locally ) ) {
+				$notices_dismissed_locally = array();
+			}
+			$notices_dismissed_locally[] = $screen;
+			update_user_option( get_current_user_id(), 'wpcom_removed_calypso_screen_dismissed_notices', $notices_dismissed_locally, true );
 		}
-		$dismissed_notices[] = $screen;
-		update_user_option( get_current_user_id(), 'wpcom_removed_calypso_screen_dismissed_notices', $dismissed_notices, true );
 	}
 	wp_die();
 }