+++ projects/packages/videopress/src/client/block-editor/blocks/video/view.ts
@@ -6,7 +6,7 @@ import domReady from '@wordpress/dom-ready';
 /**
  * Internal dependencies
  */
-import './style.scss';
+import './view.scss';
 
 /**
  * Preview on Hover effect for VideoPress videos.