+++ projects/js-packages/scan/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.4] - 2025-01-20
+### Changed
+- Updated package dependencies. [#41099]
+
 ## [0.5.3] - 2025-01-06
 ### Changed
 - Updated package dependencies. [#40798] [#40810] [#40841]
@@ -91,6 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Updated dependencies. [#39754]
 
+[0.5.4]: https://github.com/Automattic/jetpack-scan/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/Automattic/jetpack-scan/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/Automattic/jetpack-scan/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/Automattic/jetpack-scan/compare/v0.5.0...v0.5.1