+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.3.0] - 2025-01-20
+### Changed
+- Code: Use function-style exit() and die() with a default status code of 0. [#41167]
+- Move WPCOM_REST_API_Proxy_Request trait to the connection package. [#41023]
+- Updated package dependencies. [#41099]
+
+### Fixed
+- Add heartbeat deactivation on site disconnection. [#41117]
+
 ## [6.2.2] - 2025-01-06
 ### Added
 - Added tests to increase code coverage. [#39963]
@@ -1278,6 +1287,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[6.3.0]: https://github.com/Automattic/jetpack-connection/compare/v6.2.2...v6.3.0
 [6.2.2]: https://github.com/Automattic/jetpack-connection/compare/v6.2.1...v6.2.2
 [6.2.1]: https://github.com/Automattic/jetpack-connection/compare/v6.2.0...v6.2.1
 [6.2.0]: https://github.com/Automattic/jetpack-connection/compare/v6.1.1...v6.2.0